requirements.txt
altgraph==0.17.4
Brotli==1.1.0
certifi==2024.7.4
charset-normalizer==3.3.2
idna==3.7
mutagen==1.47.0
packaging==24.1
pillow==10.4.0
pycryptodomex==3.20.0
pyinstaller==6.9.0
pyinstaller-hooks-contrib==2024.7
pytube==15.0.0
requests==2.32.3
urllib3==2.2.2
websockets==12.0
yt-dlp==2024.7.25
